+++ agents/lite-rs/crates/ados-setup/web-setup/panels/index.js
@@ -8,6 +8,13 @@ export { renderCameraPanel } from "./drone/camera.js";
 export { renderSensorsPanel } from "./drone/sensors.js";
 export { renderPluginsPanel } from "./drone/plugins.js";
 
+export { renderWfbRxPanel } from "./ground/wfb_rx.js";
+export { renderMeshPanel } from "./ground/mesh.js";
+export { renderSourcesPanel } from "./ground/sources.js";
+export { renderDisplayPanel } from "./ground/display.js";
+export { renderOledButtonsPanel } from "./ground/oled_buttons.js";
+export { renderJoystickPanel } from "./ground/joystick.js";
+
 export { renderCloudPanel } from "./common/cloud.js";
 export { renderNetworkPanel } from "./common/network.js";
 export { renderServicesPanel } from "./common/services.js";